requirements.txt
coloredlogs==15.0.1
datasets==2.18.0
diffusers==0.36.0.dev0
einops==0.8.1
envlight==0.1.0
huggingface_hub==0.34.4
imageio==2.37.0
joblib==1.5.1
kaolin==0.17.0
kiui==0.2.16
kornia==0.8.1
mediapipe==0.10.21
onnxruntime-gpu==1.22.0
open3d==0.19.0
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
opencv_python_headless==4.8.0.76
PyMCubes==0.1.6
pymeshlab==2023.12.post3
pyrender==0.1.45
pytorch_lightning==2.5.1.post0
PyYAML==6.0.2
rembg==2.0.67
rich==13.4.2
six==1.17.0
smplx==0.1.28
tensorboardx==2.6.4
torch_scatter==2.1.2
tqdm==4.65.0
transformers==4.46.2
trimesh==4.6.12
triton==3.0.0
vedo==2025.5.4
xatlas==0.0.10
yacs==0.1.8
git+https://github.com/NVlabs/nvdiffrast.git